+++ dependency-check-suppressions.xml
@@ -184,6 +184,8 @@
         <cve>CVE-2021-24122</cve>
         <cve>CVE-2021-25122</cve>
         <cve>CVE-2021-25329</cve>
+        <cve>CVE-2021-30639</cve>
+        <cve>CVE-2021-33037</cve>
     </suppress>
     <suppress>
         <notes><![CDATA[
