--- src/aat/java/uk/gov/hmcts/cmc/claimstore/tests/befta/HighLevelDataSetupApp.java
@@ -1,20 +1,0 @@
-package uk.gov.hmcts.cmc.claimstore.tests.befta;
-
-import uk.gov.hmcts.befta.dse.ccd.CcdEnvironment;
-import uk.gov.hmcts.befta.dse.ccd.DataLoaderToDefinitionStore;
-
-public class HighLevelDataSetupApp extends DataLoaderToDefinitionStore {
-
-    public HighLevelDataSetupApp(CcdEnvironment dataSetupEnvironment) {
-        super(dataSetupEnvironment);
-    }
-
-    public static void main(String[] args) throws Throwable {
-        main(HighLevelDataSetupApp.class, args);
-    }
-
-    @Override
-    protected boolean shouldTolerateDataSetupFailure() {
-        return true;
-    }
-}
